+++ src/common/time_utils.c
@@ -2,25 +2,25 @@
 
 // function buffering delay() 
 // arduino uno function doesn't work well with interrupts
-void _delay(unsigned long ms){
+void _delay(uint32_t ms){
   // if arduino uno and other atmega328p chips
   // use while instad of delay, 
   // due to wrong measurement based on changed timer0
-  unsigned long t = _micros();
+  uint32_t t = _micros();
   while( _micros() - t < ms*1000 ); 
 }
 
-void _delay_us(unsigned long us)
+void _delay_us(uint32_t us)
 {
-  unsigned long t = _micros();
+  uint32_t t = _micros();
   while( _micros() - t < us ); 
 }
 
 // function buffering _micros() 
 // arduino function doesn't work well with interrupts
-static unsigned long default_micros(){
+static uint32_t default_micros(){
   return 0;
 }
 
 
-unsigned long (*_micros)(void) = default_micros;
+uint32_t (*_micros)(void) = default_micros;
